+++ client/src/components/TaskList.tsx
@@ -14,6 +14,8 @@ import CropContainer from './CropContainer'
 import EntityLink from './EntityLink'
 import EvaluationIndicator from './EvaluationIndicator'
 import TimeLimitTag from './time-limit/TimeLimitTag'
+import RelativeDateTime from './RelativeDateTime'
+import Authorized from './Authorized'
 
 const { confirm } = Modal
 
@@ -37,6 +39,18 @@ const renderTask = (props: RenderProps) => (task: Task) => {
       }
     })
 
+  const createdAtTag = (
+    <span style={{ marginRight: '1em' }}>
+      Created: <RelativeDateTime date={new Date(task.createdAt)} />
+    </span>
+  )
+
+  const updatedAtTag = (
+    <span style={{ marginRight: '1em' }}>
+      Updated: <RelativeDateTime date={new Date(task.updatedAt)} />
+    </span>
+  )
+
   const cardProps: CardProps = {
     title: (
       <>
@@ -59,19 +73,27 @@ const renderTask = (props: RenderProps) => (task: Task) => {
         ) : null}
       </>
     ),
-    extra: task.editable ? (
-      <Tooltip
-        title={task.inPool ? 'Delete from pool' : 'Remove from assignment'}
-        placement="left"
-      >
-        <Button
-          onClick={confirmDelete}
-          type="dashed"
-          shape="circle"
-          icon="delete"
-        />
-      </Tooltip>
-    ) : null,
+    extra: (
+      <>
+        <Authorized authority="ROLE_TEACHER">
+          {createdAtTag}
+          {updatedAtTag}
+        </Authorized>
+        {task.editable ? (
+          <Tooltip
+            title={task.inPool ? 'Delete from pool' : 'Remove from assignment'}
+            placement="left"
+          >
+            <Button
+              onClick={confirmDelete}
+              type="dashed"
+              shape="circle"
+              icon="delete"
+            />
+          </Tooltip>
+        ) : null}
+      </>
+    ),
     children: (
       <>
         {task.body ? (